More flexible debug modes
Sometimes you want to debug in seconds. Other times, you want the "normal" minutes. This is an explicit decision to not support seconds in non-debug mode. I don't want the tool to be, by default, any more complex than it needs to be, and there's not much point in timing oneself on tasks of sub-minute duration.